requirements.txt
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
python-dotenv==0.21.0
tomli==2.0.1
web-to-csv==0.0.0
Werkzeug==2.2.2
